Update jquery.card.js
Added possibility to trigger manual update of input field.
This can be handy when you have multiple name fields like billing / delivery / etc and they are all tied together. In this case the card details will only ever updated if the user is actually typing into the billing fields, however in many cases that won't actually happen because we use "same as delivery" which syncs the fields in between them.